refactor: move SearchColumn enums in own package #24004
