+++ db/rdbms/src/main/java/io/camunda/db/rdbms/read/service/ProcessInstanceReader.java
@@ -9,7 +9,7 @@ package io.camunda.db.rdbms.read.service;
 
 import io.camunda.db.rdbms.read.domain.ProcessInstanceDbQuery;
 import io.camunda.db.rdbms.sql.ProcessInstanceMapper;
-import io.camunda.db.rdbms.sql.ProcessInstanceMapper.ProcessInstanceSearchColumn;
+import io.camunda.db.rdbms.sql.columns.ProcessInstanceSearchColumn;
 import io.camunda.search.entities.ProcessInstanceEntity;
 import io.camunda.search.query.ProcessInstanceQuery;
 import io.camunda.search.query.SearchQueryResult;
@@ -24,7 +24,7 @@ public class ProcessInstanceReader extends AbstractEntityReader<ProcessInstanceE
   private final ProcessInstanceMapper processInstanceMapper;
 
   public ProcessInstanceReader(final ProcessInstanceMapper processInstanceMapper) {
-    super(ProcessInstanceMapper.ProcessInstanceSearchColumn::findByProperty);
+    super(ProcessInstanceSearchColumn::findByProperty);
     this.processInstanceMapper = processInstanceMapper;
   }
