+++ db/rdbms/src/main/java/io/camunda/db/rdbms/sql/DecisionRequirementsMapper.java
@@ -11,7 +11,6 @@ import io.camunda.db.rdbms.read.domain.DecisionRequirementsDbQuery;
 import io.camunda.db.rdbms.write.domain.DecisionRequirementsDbModel;
 import io.camunda.search.entities.DecisionRequirementsEntity;
 import java.util.List;
-import java.util.function.Function;
 
 public interface DecisionRequirementsMapper {
 
@@ -20,59 +19,4 @@ public interface DecisionRequirementsMapper {
   Long count(DecisionRequirementsDbQuery filter);
 
   List<DecisionRequirementsEntity> search(DecisionRequirementsDbQuery filter);
-
-  enum DecisionRequirementsSearchColumn implements SearchColumn<DecisionRequirementsEntity> {
-    DECISION_REQUIREMENTS_KEY(
-        "decisionRequirementsKey", DecisionRequirementsEntity::decisionRequirementsKey),
-    DECISION_REQUIREMENTS_ID(
-        "decisionRequirementsId", DecisionRequirementsEntity::decisionRequirementsId),
-    NAME("name", DecisionRequirementsEntity::name),
-    VERSION("version", DecisionRequirementsEntity::version),
-    TENANT_ID("tenantId", DecisionRequirementsEntity::tenantId),
-    RESOURCE_NAME("resourceName", DecisionRequirementsEntity::resourceName),
-    XML("xml", DecisionRequirementsEntity::xml);
-
-    private final String property;
-    private final Function<DecisionRequirementsEntity, Object> propertyReader;
-    private final Function<Object, Object> sortOptionConverter;
-
-    DecisionRequirementsSearchColumn(
-        final String property, final Function<DecisionRequirementsEntity, Object> propertyReader) {
-      this(property, propertyReader, Function.identity());
-    }
-
-    DecisionRequirementsSearchColumn(
-        final String property,
-        final Function<DecisionRequirementsEntity, Object> propertyReader,
-        final Function<Object, Object> sortOptionConverter) {
-      this.property = property;
-      this.propertyReader = propertyReader;
-      this.sortOptionConverter = sortOptionConverter;
-    }
-
-    @Override
-    public Object getPropertyValue(final DecisionRequirementsEntity entity) {
-      return propertyReader.apply(entity);
-    }
-
-    @Override
-    public Object convertSortOption(final Object object) {
-      if (object == null) {
-        return null;
-      }
-
-      return sortOptionConverter.apply(object);
-    }
-
-    public static DecisionRequirementsSearchColumn findByProperty(final String property) {
-      for (final DecisionRequirementsSearchColumn column :
-          DecisionRequirementsSearchColumn.values()) {
-        if (column.property.equals(property)) {
-          return column;
-        }
-      }
-
-      return null;
-    }
-  }
 }
