+++ db/rdbms/src/main/java/io/camunda/db/rdbms/sql/ProcessDefinitionMapper.java
@@ -11,7 +11,6 @@ import io.camunda.db.rdbms.read.domain.ProcessDefinitionDbQuery;
 import io.camunda.db.rdbms.write.domain.ProcessDefinitionDbModel;
 import io.camunda.search.entities.ProcessDefinitionEntity;
 import java.util.List;
-import java.util.function.Function;
 
 public interface ProcessDefinitionMapper {
 
@@ -20,58 +19,4 @@ public interface ProcessDefinitionMapper {
   Long count(ProcessDefinitionDbQuery filter);
 
   List<ProcessDefinitionEntity> search(ProcessDefinitionDbQuery filter);
-
-  enum ProcessDefinitionSearchColumn implements SearchColumn<ProcessDefinitionEntity> {
-    PROCESS_DEFINITION_KEY("processDefinitionKey", ProcessDefinitionEntity::processDefinitionKey),
-    PROCESS_DEFINITION_ID("processDefinitionId", ProcessDefinitionEntity::processDefinitionId),
-    NAME("name", ProcessDefinitionEntity::name),
-    VERSION("version", ProcessDefinitionEntity::version),
-    VERSION_TAG("versionTag", ProcessDefinitionEntity::versionTag),
-    TENANT_ID("tenantId", ProcessDefinitionEntity::tenantId),
-    FORM_ID("formId", ProcessDefinitionEntity::formId),
-    RESOURCE_NAME("resourceName", ProcessDefinitionEntity::resourceName),
-    BPMN_XML("bpmnXml", ProcessDefinitionEntity::bpmnXml);
-
-    private final String property;
-    private final Function<ProcessDefinitionEntity, Object> propertyReader;
-    private final Function<Object, Object> sortOptionConverter;
-
-    ProcessDefinitionSearchColumn(
-        final String property, final Function<ProcessDefinitionEntity, Object> propertyReader) {
-      this(property, propertyReader, Function.identity());
-    }
-
-    ProcessDefinitionSearchColumn(
-        final String property,
-        final Function<ProcessDefinitionEntity, Object> propertyReader,
-        final Function<Object, Object> sortOptionConverter) {
-      this.property = property;
-      this.propertyReader = propertyReader;
-      this.sortOptionConverter = sortOptionConverter;
-    }
-
-    @Override
-    public Object getPropertyValue(final ProcessDefinitionEntity entity) {
-      return propertyReader.apply(entity);
-    }
-
-    @Override
-    public Object convertSortOption(final Object object) {
-      if (object == null) {
-        return null;
-      }
-
-      return sortOptionConverter.apply(object);
-    }
-
-    public static ProcessDefinitionSearchColumn findByProperty(final String property) {
-      for (final ProcessDefinitionSearchColumn column : ProcessDefinitionSearchColumn.values()) {
-        if (column.property.equals(property)) {
-          return column;
-        }
-      }
-
-      return null;
-    }
-  }
 }
