+++ db/rdbms/src/main/java/io/camunda/db/rdbms/sql/columns/SearchColumn.java
@@ -5,7 +5,7 @@
  * Licensed under the Camunda License 1.0. You may not use this file
  * except in compliance with the Camunda License 1.0.
  */
-package io.camunda.db.rdbms.sql;
+package io.camunda.db.rdbms.sql.columns;
 
 public interface SearchColumn<T> {
