+++ qa/integration-tests/src/test/java/io/camunda/it/rdbms/exporter/RdbmsExporterIT.java
@@ -58,8 +58,7 @@ class RdbmsExporterIT {
 
   private final ProtocolFactory factory = new ProtocolFactory(System.nanoTime());
 
-  @Autowired
-  private RdbmsService rdbmsService;
+  @Autowired private RdbmsService rdbmsService;
 
   private RdbmsExporter exporter;
